Thanks to Evarfairy
Updated Timers.
